Davzev: Changed implementation of doCDRip and doG2Encode by splitting it into small parts like waitForServiceToFinish, checkExitValue, resetService and finishRawAudio. Now the EncodeHandler now relies on the process exitvalue when service finishes.
Also added handlers for situations that need recovery. This is when the mmbase
is reset while service is still busy, or when service just finished.
In latter situation the signal that service finished cant be received since
mmbase is rebooting.

For more info see method class comment and method comments.