keesj:bugfix methods that used mmbase classes as parameter did not work because there is no "generic" serialization for those classes.
If a client passes a remote reference back to the server, the server
will receive a stub for its own local object. to make it possible to trace
this stub to the local implementation object we now use the StubToLocalMapper.
This has not yet been stress tested. but it seams to work.

second bugfix is that the remote MMCI implementation now throws BridgeExceptions
and not RunTimeExceptions.