johannes:
This fixes a problem introduced by the getNonByteArrayFields() change in mmbase 1.6

Usually, a certain query would be 'SELECT * FROM ...'. This is now changed to
'SELECT ..., MAX, .... FROM ...' for certain builders that have the 'max' field.
Unfortunately, some databases choke on this, therefore we now prefix all fields
in the query by the tablename.

Furthermore, cleaned up the include list which got a little out of hand.