bug #6414 Javascript error on save
The only occurrence that could be found of this bug was when the referrer was passed as a parameter and contained strange encodings.
This may cause code to fail on occasions (the actual why is obscure, but related to Internet Explorer buggy behavior).
Avoiding odd encodings may help.
To aleviate the problem somewhat, I added an referrer_encoded parameter to the xsls, so at least when passing the referrer an escaped version is passed.