Fix for a hard to reproduce issue with java.rmi.NoSuchObjectException
A java.rmi.NoSuchObjectException means that the server side exported
object has been reclaimed by RMI garbage collection.

The exported object is being garbage collected in the RemoteMMCI
because no reference is kept to it.

The failure of the connection means that the server can't update the other cloud instance which results in failures